Plot a Heatmap of CNA Values

Plot a heatmap of CNA values.

cnaPlot(cna, limits = c(-1, 1), ratio = 0.5, cols = heatCols,
  x.name = "Chromosome", y.name = "Cell",
  legend.title = "Inferred CNA\n[log2 ratio]", x.hide = c("13", "18",
  "21", "Y"), orderCells = F, order.with = NULL, euclid.dist = F,
  angle = NULL, x.angle = NULL, y.angle = 0, axis.rel = 1,
  base.size = 12, axis.title.size = 12, axis.text.size = 11,
  base.col = "#073642", title = NULL, subtitle = NULL,
  caption = NULL, text.size = 12, y.hide = NULL, tile.size = 0.1,
  tile.col = NULL, legend.position = "right", legend.height = 2,
  legend.width = 0.6, legend.rel = 0.9, legend.colour = "black",
  legend.breaks = NULL, legend.labels = NULL,
  legend.justification = "top", legend.title.position = "bottom",
  legend.title.angle = NULL, legend.title.rel = 0.9)

Arguments

cna
matrix of CNA values (genes by cells).

limits
colour range. Cells >= upper limit will be the same colour, and likewise for cells <= lower limit. Default: c(-1, 1)

ratio
aspect ratio of the panel. Default: 0.7

cols
character vector of colours to use. Default: heatCols

x.name
x axis label. Default: 'Chromosome'

y.name
y axis label. Default: 'Cell'

legend.title
PARAM_DESCRIPTION, Default: 'Inferred CNA `[log2 ratio`]'

x.hide
PARAM_DESCRIPTION, Default: c("13", "18", "21", "Y")

orderCells
boolean value indicating whether cells should be ordered by hierarchical clustering. Default: F

angle
angle of axes tick labels. x.angle and y.angle inherit from angle. Default: NULL

x.angle
angle of x axis tick labels. If left, will inherit from angle. Default: NULL

y.angle
angle of y axis tick labels. If left, will inherit from angle. Default: 0

axis.rel
PARAM_DESCRIPTION, Default: 1

base.size
PARAM_DESCRIPTION, Default: 12

axis.title.size
PARAM_DESCRIPTION, Default: 12

axis.text.size
PARAM_DESCRIPTION, Default: 11

base.col
PARAM_DESCRIPTION, Default: '#073642'

title
PARAM_DESCRIPTION, Default: NULL

subtitle
PARAM_DESCRIPTION, Default: NULL

caption
PARAM_DESCRIPTION, Default: NULL

text.size
PARAM_DESCRIPTION, Default: 12

y.hide
PARAM_DESCRIPTION, Default: NULL

tile.size
PARAM_DESCRIPTION, Default: 0.1

tile.col
PARAM_DESCRIPTION, Default: NULL

legend.position
PARAM_DESCRIPTION, Default: 'right'

legend.height
PARAM_DESCRIPTION, Default: 2

legend.width
PARAM_DESCRIPTION, Default: 0.6

legend.rel
PARAM_DESCRIPTION, Default: 0.9

legend.colour
PARAM_DESCRIPTION, Default: 'black'

legend.breaks
PARAM_DESCRIPTION, Default: NULL

legend.labels
PARAM_DESCRIPTION, Default: NULL

legend.justification
PARAM_DESCRIPTION, Default: 'top'

legend.title.position
PARAM_DESCRIPTION, Default: 'bottom'

legend.title.angle
PARAM_DESCRIPTION, Default: NULL

legend.title.rel
PARAM_DESCRIPTION, Default: 0.9

Value

OUTPUT_DESCRIPTION

Details

DETAILS

See also

Examples

# NOT RUN {
if(interactive()){
 #EXAMPLE1
 }
# }