Find Clones

FUNCTION_DESCRIPTION

findClones(m, prob = 0.95, coverage = 0.8, mode.size = 10,
  clone.size = NULL, by = "chr")

Arguments

m
a matrix of genes X cells (variables X observations) containing CNA values.

prob
a numeric value >= 0 and <= 1; the minimum posterior probability required for an observation to be assigned to a mode. Default: 0.95

coverage
the fraction of observations that must have a posterior probability higher than <prob> to one of two modes in order for the distribution to qualify as bimodal. Default: 0.8

mode.size
the minimum number of observations that must be assigned to a mode in order for the distribution to qualify as bimodal. Default: 10

clone.size
PARAM_DESCRIPTION, Default: NULL

by
PARAM_DESCRIPTION, Default: 'chr'

Examples

cna = infercna(useData(), dipcells = dipcells)
# malignant cells only
cna = cna[, !colnames(cna) %in% unlist(dipcells)]
findClones(cna, by = 'chr')

#> number of iterations= 496 

#> At least one mode contains < 10 obs.

#> number of iterations= 46 

#> Success!

#> One of the variances is going to zero;  trying new starting values.
#> One of the variances is going to zero;  trying new starting values.
#> One of the variances is going to zero;  trying new starting values.
#> One of the variances is going to zero;  trying new starting values.
#> One of the variances is going to zero;  trying new starting values.
#> One of the variances is going to zero;  trying new starting values.
#> One of the variances is going to zero;  trying new starting values.
#> One of the variances is going to zero;  trying new starting values.
#> One of the variances is going to zero;  trying new starting values.
#> One of the variances is going to zero;  trying new starting values.
#> number of iterations= 76 

#> At least one mode contains < 10 obs.

#> number of iterations= 477 

#> Less than 80% of obs. could be assigned to a mode.

#> number of iterations= 6889 

#> At least one mode contains < 10 obs.

#> number of iterations= 6377 

#> At least one mode contains < 10 obs.

#> number of iterations= 20 

#> At least one mode contains < 10 obs.

#> number of iterations= 1036 

#> At least one mode contains < 10 obs.

#> number of iterations= 312 

#> At least one mode contains < 10 obs.

#> number of iterations= 65 

#> At least one mode contains < 10 obs.

#> number of iterations= 1522 

#> At least one mode contains < 10 obs.

#> number of iterations= 263 

#> At least one mode contains < 10 obs.

#> number of iterations= 318 

#> At least one mode contains < 10 obs.

#> number of iterations= 454 

#> At least one mode contains < 10 obs.

#> number of iterations= 157 

#> At least one mode contains < 10 obs.

#> number of iterations= 724 

#> At least one mode contains < 10 obs.

#> number of iterations= 1095 

#> At least one mode contains < 10 obs.

#> number of iterations= 1381 

#> At least one mode contains < 10 obs.

#> number of iterations= 163 

#> At least one mode contains < 10 obs.

#> number of iterations= 493 

#> At least one mode contains < 10 obs.

#> number of iterations= 1513 

#> At least one mode contains < 10 obs.

#> number of iterations= 262 

#> At least one mode contains < 10 obs.

#> number of iterations= 679 

#> At least one mode contains < 10 obs.

#> number of iterations= 161 

#> At least one mode contains < 10 obs.

#> $`2.a`
#>  [1] "MGH125-P11-A04" "MGH125-P11-A08" "MGH125-P11-B02" "MGH125-P11-C03"
#>  [5] "MGH125-P11-C08" "MGH125-P11-C12" "MGH125-P11-E09" "MGH125-P11-G02"
#>  [9] "MGH125-P11-G07" "MGH125-P2-A01"  "MGH125-P2-A06"  "MGH125-P2-A10" 
#> [13] "MGH125-P2-A11"  "MGH125-P2-B05"  "MGH125-P2-B07"  "MGH125-P2-B09" 
#> [17] "MGH125-P2-B10"  "MGH125-P2-B12"  "MGH125-P2-C01"  "MGH125-P2-C03" 
#> [21] "MGH125-P2-C09"  "MGH125-P2-C12"  "MGH125-P2-D03"  "MGH125-P2-D11" 
#> [25] "MGH125-P2-E07"  "MGH125-P2-F01"  "MGH125-P2-F05"  "MGH125-P2-F11" 
#> [29] "MGH125-P2-F12"  "MGH125-P2-G12"  "MGH125-P2-H01"  "MGH125-P3-C06" 
#> [33] "MGH125-P3-C10"  "MGH125-P3-C12"  "MGH125-P3-D02"  "MGH125-P3-D06" 
#> [37] "MGH125-P3-D08"  "MGH125-P3-E01"  "MGH125-P3-E09"  "MGH125-P3-F05" 
#> [41] "MGH125-P3-G02"  "MGH125-P3-G09"  "MGH125-P3-H03"  "MGH125-P3-H04" 
#> [45] "MGH125-P3-H10"  "MGH125-P3-H11"  "MGH125-P4-A03"  "MGH125-P4-A09" 
#> [49] "MGH125-P4-B01"  "MGH125-P4-B04"  "MGH125-P4-B06"  "MGH125-P4-B08" 
#> [53] "MGH125-P4-C01"  "MGH125-P4-C02"  "MGH125-P4-C04"  "MGH125-P4-C11" 
#> [57] "MGH125-P4-D03"  "MGH125-P4-D12"  "MGH125-P4-E02"  "MGH125-P4-E09" 
#> [61] "MGH125-P4-F02"  "MGH125-P4-F03"  "MGH125-P4-F10"  "MGH125-P4-G02" 
#> [65] "MGH125-P4-G07"  "MGH125-P4-G10"  "MGH125-P4-H06"  "MGH125-P7-A04" 
#> [69] "MGH125-P7-B01"  "MGH125-P7-B02"  "MGH125-P7-B03"  "MGH125-P7-B07" 
#> [73] "MGH125-P7-B09"  "MGH125-P7-B10"  "MGH125-P7-C02"  "MGH125-P7-C06" 
#> [77] "MGH125-P7-D06"  "MGH125-P7-D10"  "MGH125-P7-E05"  "MGH125-P7-E12" 
#> [81] "MGH125-P7-F03"  "MGH125-P7-G05"  "MGH125-P7-G07"  "MGH125-P7-H04" 
#> [85] "MGH125-P7-H09"  "MGH125-P7-H12" 
#> 
#> $`2.b`
#>   [1] "MGH125-P2-A02" "MGH125-P2-A03" "MGH125-P2-A05" "MGH125-P2-A09"
#>   [5] "MGH125-P2-B01" "MGH125-P2-B03" "MGH125-P2-B04" "MGH125-P2-B06"
#>   [9] "MGH125-P2-B08" "MGH125-P2-C02" "MGH125-P2-C05" "MGH125-P2-C06"
#>  [13] "MGH125-P2-C07" "MGH125-P2-C08" "MGH125-P2-C10" "MGH125-P2-C11"
#>  [17] "MGH125-P2-D02" "MGH125-P2-D04" "MGH125-P2-D05" "MGH125-P2-D07"
#>  [21] "MGH125-P2-D08" "MGH125-P2-D09" "MGH125-P2-D10" "MGH125-P2-D12"
#>  [25] "MGH125-P2-E01" "MGH125-P2-E02" "MGH125-P2-E04" "MGH125-P2-E05"
#>  [29] "MGH125-P2-E06" "MGH125-P2-E09" "MGH125-P2-E10" "MGH125-P2-E12"
#>  [33] "MGH125-P2-F02" "MGH125-P2-F03" "MGH125-P2-F04" "MGH125-P2-F06"
#>  [37] "MGH125-P2-F07" "MGH125-P2-F10" "MGH125-P2-G02" "MGH125-P2-G03"
#>  [41] "MGH125-P2-G06" "MGH125-P2-G07" "MGH125-P2-G08" "MGH125-P2-G09"
#>  [45] "MGH125-P2-G10" "MGH125-P2-G11" "MGH125-P2-H02" "MGH125-P2-H03"
#>  [49] "MGH125-P2-H04" "MGH125-P2-H07" "MGH125-P2-H08" "MGH125-P2-H11"
#>  [53] "MGH125-P2-H12" "MGH125-P3-A02" "MGH125-P3-A12" "MGH125-P3-B02"
#>  [57] "MGH125-P3-B04" "MGH125-P3-B05" "MGH125-P3-B06" "MGH125-P3-B07"
#>  [61] "MGH125-P3-B08" "MGH125-P3-B09" "MGH125-P3-B10" "MGH125-P3-B11"
#>  [65] "MGH125-P3-B12" "MGH125-P3-C01" "MGH125-P3-C02" "MGH125-P3-C03"
#>  [69] "MGH125-P3-C05" "MGH125-P3-C07" "MGH125-P3-C08" "MGH125-P3-C09"
#>  [73] "MGH125-P3-D01" "MGH125-P3-D04" "MGH125-P3-D05" "MGH125-P3-D07"
#>  [77] "MGH125-P3-D09" "MGH125-P3-D10" "MGH125-P3-D11" "MGH125-P3-D12"
#>  [81] "MGH125-P3-E02" "MGH125-P3-E03" "MGH125-P3-E04" "MGH125-P3-E05"
#>  [85] "MGH125-P3-E06" "MGH125-P3-E07" "MGH125-P3-E08" "MGH125-P3-E10"
#>  [89] "MGH125-P3-E11" "MGH125-P3-E12" "MGH125-P3-F02" "MGH125-P3-F03"
#>  [93] "MGH125-P3-F04" "MGH125-P3-F06" "MGH125-P3-F08" "MGH125-P3-F10"
#>  [97] "MGH125-P3-F11" "MGH125-P3-F12" "MGH125-P3-G01" "MGH125-P3-G03"
#> [101] "MGH125-P3-G04" "MGH125-P3-G05" "MGH125-P3-G06" "MGH125-P3-G07"
#> [105] "MGH125-P3-G08" "MGH125-P3-G10" "MGH125-P3-G11" "MGH125-P3-G12"
#> [109] "MGH125-P3-H01" "MGH125-P3-H05" "MGH125-P3-H07" "MGH125-P3-H08"
#> [113] "MGH125-P3-H09" "MGH125-P3-H12" "MGH125-P4-A01" "MGH125-P4-A02"
#> [117] "MGH125-P4-A04" "MGH125-P4-A05" "MGH125-P4-A06" "MGH125-P4-A07"
#> [121] "MGH125-P4-A08" "MGH125-P4-A10" "MGH125-P4-A11" "MGH125-P4-A12"
#> [125] "MGH125-P4-B02" "MGH125-P4-B03" "MGH125-P4-B05" "MGH125-P4-B07"
#> [129] "MGH125-P4-B09" "MGH125-P4-B10" "MGH125-P4-B11" "MGH125-P4-B12"
#> [133] "MGH125-P4-C03" "MGH125-P4-C05" "MGH125-P4-C06" "MGH125-P4-C07"
#> [137] "MGH125-P4-C08" "MGH125-P4-C09" "MGH125-P4-C10" "MGH125-P4-C12"
#> [141] "MGH125-P4-D01" "MGH125-P4-D02" "MGH125-P4-D04" "MGH125-P4-D05"
#> [145] "MGH125-P4-D06" "MGH125-P4-D07" "MGH125-P4-D08" "MGH125-P4-D09"
#> [149] "MGH125-P4-D10" "MGH125-P4-E01" "MGH125-P4-E03" "MGH125-P4-E04"
#> [153] "MGH125-P4-E06" "MGH125-P4-E07" "MGH125-P4-E08" "MGH125-P4-E10"
#> [157] "MGH125-P4-E11" "MGH125-P4-E12" "MGH125-P4-F01" "MGH125-P4-F04"
#> [161] "MGH125-P4-F05" "MGH125-P4-F06" "MGH125-P4-F07" "MGH125-P4-F08"
#> [165] "MGH125-P4-F09" "MGH125-P4-F11" "MGH125-P4-F12" "MGH125-P4-G03"
#> [169] "MGH125-P4-G05" "MGH125-P4-G06" "MGH125-P4-G08" "MGH125-P4-G09"
#> [173] "MGH125-P4-G11" "MGH125-P4-G12" "MGH125-P4-H01" "MGH125-P4-H02"
#> [177] "MGH125-P4-H04" "MGH125-P7-A01" "MGH125-P7-A02" "MGH125-P7-A03"
#> [181] "MGH125-P7-A05" "MGH125-P7-A07" "MGH125-P7-A08" "MGH125-P7-A09"
#> [185] "MGH125-P7-A10" "MGH125-P7-A11" "MGH125-P7-B05" "MGH125-P7-B08"
#> [189] "MGH125-P7-C01" "MGH125-P7-C03" "MGH125-P7-C04" "MGH125-P7-C05"
#> [193] "MGH125-P7-C07" "MGH125-P7-C08" "MGH125-P7-C09" "MGH125-P7-C10"
#> [197] "MGH125-P7-C11" "MGH125-P7-C12" "MGH125-P7-D01" "MGH125-P7-D02"
#> [201] "MGH125-P7-D04" "MGH125-P7-D05" "MGH125-P7-D07" "MGH125-P7-D08"
#> [205] "MGH125-P7-D09" "MGH125-P7-D11" "MGH125-P7-D12" "MGH125-P7-E01"
#> [209] "MGH125-P7-E02" "MGH125-P7-E03" "MGH125-P7-E04" "MGH125-P7-E06"
#> [213] "MGH125-P7-E07" "MGH125-P7-E08" "MGH125-P7-E09" "MGH125-P7-E10"
#> [217] "MGH125-P7-F01" "MGH125-P7-F02" "MGH125-P7-F04" "MGH125-P7-F05"
#> [221] "MGH125-P7-F06" "MGH125-P7-F07" "MGH125-P7-F08" "MGH125-P7-F09"
#> [225] "MGH125-P7-F10" "MGH125-P7-F11" "MGH125-P7-G01" "MGH125-P7-G02"
#> [229] "MGH125-P7-G03" "MGH125-P7-G04" "MGH125-P7-G06" "MGH125-P7-G09"
#> [233] "MGH125-P7-G10" "MGH125-P7-G11" "MGH125-P7-G12" "MGH125-P7-H02"
#> [237] "MGH125-P7-H03" "MGH125-P7-H05" "MGH125-P7-H06" "MGH125-P7-H10"
#> [241] "MGH125-P7-H11"
#> 

findClones(cna, by = 'arm')

#> One of the variances is going to zero;  trying new starting values.
#> One of the variances is going to zero;  trying new starting values.
#> number of iterations= 121 

#> At least one mode contains < 10 obs.

#> number of iterations= 343 

#> At least one mode contains < 10 obs.

#> number of iterations= 151 

#> Less than 80% of obs. could be assigned to a mode.

#> number of iterations= 3 

#> At least one mode contains < 10 obs.

#> number of iterations= 164 

#> At least one mode contains < 10 obs.

#> number of iterations= 144 

#> At least one mode contains < 10 obs.

#> number of iterations= 31 

#> At least one mode contains < 10 obs.

#> number of iterations= 315 

#> At least one mode contains < 10 obs.

#> number of iterations= 320 

#> At least one mode contains < 10 obs.

#> number of iterations= 413 

#> At least one mode contains < 10 obs.

#> number of iterations= 697 

#> At least one mode contains < 10 obs.

#> number of iterations= 382 

#> At least one mode contains < 10 obs.

#> number of iterations= 18 

#> At least one mode contains < 10 obs.

#> number of iterations= 15 

#> At least one mode contains < 10 obs.

#> number of iterations= 249 

#> At least one mode contains < 10 obs.

#> number of iterations= 6624 

#> At least one mode contains < 10 obs.

#> number of iterations= 5698 

#> At least one mode contains < 10 obs.

#> number of iterations= 107 

#> At least one mode contains < 10 obs.

#> number of iterations= 396 

#> At least one mode contains < 10 obs.

#> number of iterations= 146 

#> At least one mode contains < 10 obs.

#> number of iterations= 293 

#> At least one mode contains < 10 obs.

#> One of the variances is going to zero;  trying new starting values.
#> One of the variances is going to zero;  trying new starting values.
#> One of the variances is going to zero;  trying new starting values.
#> One of the variances is going to zero;  trying new starting values.
#> One of the variances is going to zero;  trying new starting values.
#> One of the variances is going to zero;  trying new starting values.
#> One of the variances is going to zero;  trying new starting values.
#> One of the variances is going to zero;  trying new starting values.
#> One of the variances is going to zero;  trying new starting values.
#> One of the variances is going to zero;  trying new starting values.
#> One of the variances is going to zero;  trying new starting values.
#> One of the variances is going to zero;  trying new starting values.
#> One of the variances is going to zero;  trying new starting values.
#> One of the variances is going to zero;  trying new starting values.
#> One of the variances is going to zero;  trying new starting values.
#> One of the variances is going to zero;  trying new starting values.
#> One of the variances is going to zero;  trying new starting values.
#> One of the variances is going to zero;  trying new starting values.
#> One of the variances is going to zero;  trying new starting values.
#> One of the variances is going to zero;  trying new starting values.
#> number of iterations= 131 

#> At least one mode contains < 10 obs.

#> number of iterations= 2823 

#> At least one mode contains < 10 obs.

#> number of iterations= 82 

#> At least one mode contains < 10 obs.

#> number of iterations= 264 

#> At least one mode contains < 10 obs.

#> number of iterations= 631 

#> At least one mode contains < 10 obs.

#> number of iterations= 193 

#> At least one mode contains < 10 obs.

#> number of iterations= 437 

#> Less than 80% of obs. could be assigned to a mode.

#> number of iterations= 515 

#> At least one mode contains < 10 obs.

#> number of iterations= 2200 

#> At least one mode contains < 10 obs.

#> number of iterations= 701 

#> At least one mode contains < 10 obs.

#> number of iterations= 53 

#> At least one mode contains < 10 obs.

#> WARNING! NOT CONVERGENT! 
#> number of iterations= 10000 

#> At least one mode contains < 10 obs.

#> number of iterations= 666 

#> At least one mode contains < 10 obs.

#> number of iterations= 529 

#> At least one mode contains < 10 obs.

#> number of iterations= 693 

#> At least one mode contains < 10 obs.

#> WARNING! NOT CONVERGENT! 
#> number of iterations= 10000 

#> At least one mode contains < 10 obs.

#> number of iterations= 383 

#> At least one mode contains < 10 obs.

#> number of iterations= 1625 

#> At least one mode contains < 10 obs.

#> number of iterations= 1186 

#> At least one mode contains < 10 obs.

#> number of iterations= 1560 

#> At least one mode contains < 10 obs.

#> number of iterations= 180 

#> At least one mode contains < 10 obs.

#> number of iterations= 200 

#> At least one mode contains < 10 obs.

#> number of iterations= 472 

#> At least one mode contains < 10 obs.

#> Error in split.default(names(modes), modes): first argument must be a vector