Test Distribution For Modality

Test the modality (gaussian) of a distribution. Default tests for uni-, bi- and tri- modalities (argument <modes> = 2:3).

modality(x, modes = 2:3, prob = 0.95, coverage = 0.8, size = 10,
  ...)

Arguments

x
a named numeric vector of cells/observations or a matrix of genes X cells (variables X observations). If the latter, the column means are first computed.

modes
the modes to test; a numeric value or a numeric vector of modes to test. e.g. to test just for bimodality, modes = 2. Default: 2:3

prob
a numeric value >= 0 and <= 1; the minimeanm posterior probability required for an observation to be assigned to a mode. Default: 0.95

coverage
the fraction of observations that meanst have a posterior probability higher than <prob> to one of two modes in order for the distribution to qualify as bimodal. Default: 0.8

size
the minimeanm number of observations that meanst be assigned to a mode in order for the distribution to qualify as bimodal. Default: 10

...
other arguments passed to fitBimodal

Value

vector of boolean values of length equal to length(modes) that were tested

Examples

x = c(rnorm(50, mean = 1), rnorm(50, mean = 10))
modality(x, modes = 2)

#> Error in (function (x, prob = 0.95, coverage = 0.8, size = 10, assign = FALSE,     boolean = FALSE, verbose = TRUE, maxit = 5000, maxrestarts = 100,     bySampling = FALSE, nsamp = 200, ...) {    if (bySampling) {        out = .fitBimodalBySampling(x = x, tries = nsamp, prob = prob,             coverage = coverage, size = size, verbose = verbose,             maxit = maxit, maxrestarts = maxrestarts, ...)        return(out)    }    if (!is.null(dim(x)))         x = colMeans(x)    if (length(x) < size * 2) {        stop("Number of observations is too small for 2 modes >= ",             size)    }    obj = nor1mix::norMixEM(x, m = 2, maxit = maxit)    if (isFALSE(attr(obj, "converged"))) {        if (verbose)             message("No bimodal distribution found.")        return(FALSE)    }    result = nor1mix::estep.nm(x, obj)    passed = result >= prob    failed = is.null(dim(passed)) || ncol(passed) != 2 || nrow(passed) !=         length(x)    if (failed) {        if (verbose)             message("No bimodal distribution found.")        return(FALSE)    }    if (any(colSums(passed) < size)) {        if (verbose)             message("At least one mode contains < ", size, " obs.")        return(FALSE)    }    observed = (sum(colSums(passed)))/length(x)    if (observed < coverage) {        if (verbose)             message("Less than ", coverage * 100, "% of obs. could be assigned to a mode.")        return(FALSE)    }    if (verbose)         message("Success!")    if (assign) {        L = sapply(as.data.frame(passed), function(col) names(x)[which(col)],             simplify = F)        return(stats::setNames(L, letters[1:length(L)]))    }    if (boolean)         return(TRUE)    result})(x = c(-0.400043516721755, 1.25531705484526, -1.43726361121953, 0.994428713253839, 1.62155272141521, 2.14841160602606, -0.821817660976626, 0.752674697926476, 0.755800393221617, 0.717294551185535, 0.446300616311279, 1.62898204203601, 3.06502489535922, -0.630989402082232, 1.5124269498518, -0.863011492068328, 0.477987485254546, 0.94739809004612, 1.54299634266114, 0.0859251727400716, 1.46815442045053, 1.36295125586499, -0.304543545034777, 1.73777632125505, 2.88850492923455, 0.902554895591794, 0.064152646499322, 0.984049688749462, 0.173211046266557, -0.512399651262802, 1.93536318980109, 1.17648861056246, 1.2436854647462, 2.6235488834436, 1.11203808286976, 0.866002986937445, -0.910087467722212, 0.720762758415155, 0.68655402174211, 2.06730787876151, 1.07003485029664, 0.360876676114336, 0.950035100951855, 0.748516556749786, 1.44479711579273, 3.75541757533686, 1.04653138044419, 1.57770906930663, 1.11819487440968, -0.911720491028202, 10.8620864818386, 9.75676326036652, 9.79391280521842, 10.0191775922091, 10.0295607543198, 10.5498275418017, 7.72588514310706, 12.6825571835563, 9.63877874450628, 10.21335574971, 11.0743458822597, 9.3349117514545, 11.1139524189692, 9.75410358832085, 8.82243669105672, 9.02414938384014, 11.0650573200272, 10.1316706347729, 10.488628809201, 8.30054943215431, 8.52926369362257, 10.2841503436269, 11.3373204132356, 10.2366962832865, 11.3182933840346, 10.5239097875012, 10.6067480465571, 9.89006432785513, 10.172181715411, 9.90967271299934, 11.9243433413578, 11.2983927588309, 10.7487912675368, 10.5562243291043, 9.45174273626114, 11.1105348925818, 7.3876656671157, 9.84430622376995, 10.4338897899172, 9.61804888818383, 10.4241875745475, 11.0631019958488, 11.0487126197557, 9.96189710518687, 10.4861489195152, 11.6728826108325, 9.64563883644356, 10.9463478864665, 11.3168263560442, 9.70335997540817), size = 10, coverage = 0.8, prob = 0.95, boolean = TRUE,     m = 2): argument 6 matches multiple formal arguments

modality(x, modes = 2:3)

#> Error in (function (x, prob = 0.95, coverage = 0.8, size = 10, assign = FALSE,     boolean = FALSE, verbose = TRUE, maxit = 5000, maxrestarts = 100,     bySampling = FALSE, nsamp = 200, ...) {    if (bySampling) {        out = .fitBimodalBySampling(x = x, tries = nsamp, prob = prob,             coverage = coverage, size = size, verbose = verbose,             maxit = maxit, maxrestarts = maxrestarts, ...)        return(out)    }    if (!is.null(dim(x)))         x = colMeans(x)    if (length(x) < size * 2) {        stop("Number of observations is too small for 2 modes >= ",             size)    }    obj = nor1mix::norMixEM(x, m = 2, maxit = maxit)    if (isFALSE(attr(obj, "converged"))) {        if (verbose)             message("No bimodal distribution found.")        return(FALSE)    }    result = nor1mix::estep.nm(x, obj)    passed = result >= prob    failed = is.null(dim(passed)) || ncol(passed) != 2 || nrow(passed) !=         length(x)    if (failed) {        if (verbose)             message("No bimodal distribution found.")        return(FALSE)    }    if (any(colSums(passed) < size)) {        if (verbose)             message("At least one mode contains < ", size, " obs.")        return(FALSE)    }    observed = (sum(colSums(passed)))/length(x)    if (observed < coverage) {        if (verbose)             message("Less than ", coverage * 100, "% of obs. could be assigned to a mode.")        return(FALSE)    }    if (verbose)         message("Success!")    if (assign) {        L = sapply(as.data.frame(passed), function(col) names(x)[which(col)],             simplify = F)        return(stats::setNames(L, letters[1:length(L)]))    }    if (boolean)         return(TRUE)    result})(x = c(-0.400043516721755, 1.25531705484526, -1.43726361121953, 0.994428713253839, 1.62155272141521, 2.14841160602606, -0.821817660976626, 0.752674697926476, 0.755800393221617, 0.717294551185535, 0.446300616311279, 1.62898204203601, 3.06502489535922, -0.630989402082232, 1.5124269498518, -0.863011492068328, 0.477987485254546, 0.94739809004612, 1.54299634266114, 0.0859251727400716, 1.46815442045053, 1.36295125586499, -0.304543545034777, 1.73777632125505, 2.88850492923455, 0.902554895591794, 0.064152646499322, 0.984049688749462, 0.173211046266557, -0.512399651262802, 1.93536318980109, 1.17648861056246, 1.2436854647462, 2.6235488834436, 1.11203808286976, 0.866002986937445, -0.910087467722212, 0.720762758415155, 0.68655402174211, 2.06730787876151, 1.07003485029664, 0.360876676114336, 0.950035100951855, 0.748516556749786, 1.44479711579273, 3.75541757533686, 1.04653138044419, 1.57770906930663, 1.11819487440968, -0.911720491028202, 10.8620864818386, 9.75676326036652, 9.79391280521842, 10.0191775922091, 10.0295607543198, 10.5498275418017, 7.72588514310706, 12.6825571835563, 9.63877874450628, 10.21335574971, 11.0743458822597, 9.3349117514545, 11.1139524189692, 9.75410358832085, 8.82243669105672, 9.02414938384014, 11.0650573200272, 10.1316706347729, 10.488628809201, 8.30054943215431, 8.52926369362257, 10.2841503436269, 11.3373204132356, 10.2366962832865, 11.3182933840346, 10.5239097875012, 10.6067480465571, 9.89006432785513, 10.172181715411, 9.90967271299934, 11.9243433413578, 11.2983927588309, 10.7487912675368, 10.5562243291043, 9.45174273626114, 11.1105348925818, 7.3876656671157, 9.84430622376995, 10.4338897899172, 9.61804888818383, 10.4241875745475, 11.0631019958488, 11.0487126197557, 9.96189710518687, 10.4861489195152, 11.6728826108325, 9.64563883644356, 10.9463478864665, 11.3168263560442, 9.70335997540817), size = 10, coverage = 0.8, prob = 0.95, boolean = TRUE,     m = 2L): argument 6 matches multiple formal arguments